fix(TDI-39056): extend types support & handle Nullable columns in dynamodbOutput
Support of Talend Types precisions :
Date type are formatted to String using the user pattern and stored to dynamodb as a String (the supported dynamodb format for date)

Nullable columns are submited to dynamodb as Null(DynamoDB type)
If column is of type String and is nullable. If the column value is empty we also submit it as Null (DynamoDB type).

If the column are non Nullable and the column value are Null (or Empty in case of String type) we ignore them in the insertion.
DynamoDB expect Attribute values to not be null. String and Binary type attributes must have lengths greater than zero.