fix(TBD-5781): handle tKafkaOutput exceptions
Previously we didn't effectively handle the kafka producer exceptions
because async behaviour of the producer's send method. So we catched it
by creating a Callback then pass back and throw it up in our main thread.

Add Die_on_error checkbox for tKafkaOutput, to stop job for some
uncritical exceptions